Future-proof support for more than 256 tiles
BG tile VRAM has a top and bottom row, and a left and right column (in BGB's viewer).
Internally, tile IDs 00-7F are on the bottom, 80-FF on top, and the attributes' bank bit picks a left/right column.
In pokecrystal's data, tile IDs 00-7F are on the bottom-left, 80-FF on the bottom-right, and the top is inaccessible.
Now, the "extra" bit picks a bottom-top row.